Struct dns::Header

pub struct Header {
    pub id: u16,
    pub(crate) flags: u16,
    pub qdcount: u16,
    pub ancount: u16,
    pub nscount: u16,
    pub arcount: u16,
}

Fields
id: u16

A 16 bit identifier assigned by the program that generates any kind of query. This identifier is copied the corresponding reply and can be used by the requester to match up replies to outstanding queries.

flags: u16

See the Header::set_flags() method for more information.

qdcount: u16

An unsigned 16 bit integer specifying the number of entries in the question section.

ancount: u16

an unsigned 16 bit integer specifying the number of resource records in the answer section.

nscount: u16

an unsigned 16 bit integer specifying the number of name server resource records in the authority records section.

arcount: u16

an unsigned 16 bit integer specifying the number of resource records in the additional records section.

Implementations

impl Header

pub fn new() -> Header

Creates a new DnsOption

Example

use dns::Header;

let options = Header::new();

Note

This method creates a new DnsOption with the following default values:

id: 0

flags: 0

question: Vec::new()

qdcount: 0

ancount: 0

nscount: 0

arcount: 0

qtype: 0

qclass: 0

pub fn set_id(&mut self, id: u16)

Sets the id

Arguments

takes 1 u16 argument.

Example

use dns::Header;

let mut options = Header::new();
options.set_id(1234);

pub fn set_flags(&mut self, flags: u16) -> Result<(), Box<dyn Error>>

Sets the request flags

Arguments

takes 1 u16 argument with the following format:

1 bit: QR (Query/Response)

A one bit field that specifies whether this message is a query (0), or a response (1).

4 bits: Opcode A four bit field that specifies kind of query in this message. This value is set by the originator of a query and copied into the response. The values are:

0: a standard query (QUERY)

1: an inverse query (IQUERY)

2: a server status request (STATUS)

3-15: reserved for future use

DNS Header Flags (source)

1 bit: AA (Authoritative Answer)

Authoritative Answer - this bit is valid in responses, and specifies that the responding name server is an authority for the domain name in question section. Note that the contents of the answer section may have multiple owner names because of aliases.

1 bit: TC (Truncated)

TrunCation - specifies that this message was truncated due to length greater than that permitted on the transmission channel.

1 bit: RD (Recursion Desired)

Recursion Desired - this bit may be set in a query and is copied into the response. If RD is set, it directs the name server to pursue the query recursively. Recursive query support is optional.

1 bit: RA (Recursion Available)

Recursion Available - this be is set or cleared in a response, and denotes whether recursive query support is available in the name server.

3 bits: Z (Reserved)

Reserved for future use. Must be zero in all queries and responses.

4 bits: RCODE (Response Code)

Response code - this 4 bit field is set as part of responses. The values have the following interpretation:

0: No error condition

1: Format error - The name server was unable to interpret the query.

2: Server failure - The name server was unable to process this query due to a problem with the name server.

3: Name Error - Meaningful only for responses from an authoritative name server, this code signifies that the domain name referenced in the query does not exist.

4: Not Implemented - The name server does not support the requested kind of query.

5: Refused - The name server refuses to perform the specified operation for policy reasons. For example, a name server may not wish to provide the information to the particular requester, or a name server may not wish to perform a particular operation (e.g., zone transfer) for particular data.

6-15: Reserved for future use.

Example

use dns::Header;

let mut options = Header::new();
options.set_flags(0b0000_0000_0000_0000);